removing DataProviderSource.java
